+++ examples/main.tf
@@ -1,4 +1,0 @@
-resource "azurerm_resource_group" "this" {
-  name     = uuid()
-  location = "westeurope"
-}