+++ CHANGELOG.md
@@ -9,6 +9,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.0.0] - 2022-07-25
 
+### Features
+
+- Initial Commit
+- Create ```Virtual Network Setup with an Azure Firewall``` with:
+  - Virutal Network (VNET)
+  - Azure Firewall
+  - Private DNS Zone
+  - Network Security Group
+  - Route Table
+  - Bastion Host
+
+
 [Unreleased]: https://github.com/patrickhayo/azr-tf-example-firewall-network/compare/1.0.0...HEAD
 
 [1.0.0]: https://github.com/patrickhayo/azr-tf-example-firewall-network/compare/5161d3d8f6954b79125f61ee13ba01ed9bbdd47f...1.0.0